+++ src/app/employee/employee.component.ts
@@ -3,19 +3,24 @@ import { EmployeeService } from '../services/employee.service';
 import { Employee } from '../models/employee.model';
 import { DepartmentService } from '../services/department.service';
 import { Department } from '../models/department.model';
+import { MatDialog } from '@angular/material';
+import { ConfirmDialogComponent } from '../confirm-dialog.component';
 
 @Component({
   selector: 'app-employee',
   templateUrl: './employee.component.html',
   styleUrls: ['./employee.component.css']
 })
+
 export class EmployeeComponent implements OnInit {
   employees: Employee[] = [];
   dpts: Department[] = []
   formData: Employee = { id: 0, name: '', position: '', departmentId: 0, phone: null, salary: null };
   displayedColumns: string[] = ['name', 'position', 'departmentId', 'phone', 'salary', 'actions'];
 
-  constructor(private empService: EmployeeService, private dptService: DepartmentService) { }
+  constructor(private empService: EmployeeService, private dptService: DepartmentService,
+    private dialog: MatDialog
+  ) { }
 
   ngOnInit() {
     this.loadEmployees();
@@ -33,7 +38,7 @@ export class EmployeeComponent implements OnInit {
 
   onSubmit() {
     if (this.formData.id === 0) {
-      const maxId = this.employees.length ? Math.max(...this.employees.map(d => d.id)): 0;
+      const maxId = this.employees.length ? Math.max(...this.employees.map(d => d.id)) : 0;
       this.formData.id = maxId + 1;
       this.empService.addEmployee(this.formData).subscribe(() => this.loadEmployees());
     } else {
@@ -47,7 +52,12 @@ export class EmployeeComponent implements OnInit {
   }
 
   delete(id: number) {
-    this.empService.deleteEmployee(id).subscribe(() => this.loadEmployees());
+    const dialogRef = this.dialog.open(ConfirmDialogComponent)
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.empService.deleteEmployee(id).subscribe(() => this.loadEmployees());
+      }
+    });
   }
 
   resetForm() {
